Count URL length as Mastodon
Mastodon always counts a URL as 23 characters, no matter if it's longer
or shorter, so when we truncated at 500 characters it could cut off a
URL and make it invalid, even if Mastodon would have accepted it.

Instead, if a URL is at the 500 character barrier, then truncate
*before* the URL to retain the existing behavior and no broken URLs.

The test does not use the constants because they are intended to verify
that we don't accidentally change anything, so that's why we use the
magic values there, but documented.